Remove unnecessary integer cast

+++ l7_update_perf.c
@@ -270,7 +270,7 @@ int gauss_dist_int(double mean, double stdev) {
     // an additional number can be generated in the
     // same distribution using the alternate form
     // ((r*sin(theta)) * stdev) + mean
-    return (int)round(((r*cos(theta)) * stdev) + mean);
+    return round(((r*cos(theta)) * stdev) + mean);
 }
 
 // casts two void pointers (va, vb) into doubles (a, b)
